autoprint 2026/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
requests==2.32.3
python-liquid==1.12.1
pydantic==2.7.4
python-multipart==0.0.9
PyYAML==6.0.1
pywin32==306; platform_system == "Windows"
qrcode==7.4.2
Pillow==10.4.0
